lib/src/image_qrcode/requirements.txt
Pillow
pandas
psd-tools
qrcode
